Implement async IO for CPU
We now have input/output functions return None to signal Halt.

This way we can launch a thread running a CPU (via `run_async`), and
have it run indefinitely until the other thread communicating with the
CPU decides to close the communication channels. Then once the channels
are closed, the CPU is stopped and the thread is cleaned up.

See Day 15 for an example.